Update README.md
I don't think making the passwords to video-recording of pioneer Q&A sessions on a public repo is OK, without asking first for explicit permission.   
For two reasons:
 (i) it violates privacy of attendees
(ii) it is pointless to have the passcode.  
If you already asked and got official confirmation/permission then please ignore this and add a confirmation note to the repo.